scripts/ingestion/requirements.txt
psycopg2-binary>=2.9,<3
google-api-python-client>=2.100,<3
google-auth>=2.20,<3
google-cloud-resource-manager>=1.12,<2
boto3>=1.34,<2
requests>=2.31,<3
APScheduler>=3.10,<4
python-dotenv>=1.0,<2
